Move setting default displayed routes to useRouteLabelsQueryParam
