+++ ui/src/components/routes-and-lines/line-details/LineDetailsPage.tsx
@@ -1,7 +1,5 @@
 import { DateTime } from 'luxon';
-import { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
-import { pipe, uniq } from 'remeda';
 import { LineAllFieldsFragment } from '../../../generated/graphql';
 import {
   useAppDispatch,
@@ -34,10 +32,11 @@ export const LineDetailsPage = (): JSX.Element => {
 
   const dispatch = useAppDispatch();
   const { addMapOpenQueryParameter } = useMapQueryParams();
-  const { displayedRouteLabels, setDisplayedRoutesToUrl } =
-    useRouteLabelsQueryParam();
 
   const { line } = useGetLineDetails();
+
+  const { displayedRouteLabels } =
+    useRouteLabelsQueryParam(line);
 
   const createRoute = (routeLine: LineAllFieldsFragment) => {
     dispatch(resetMapRouteEditorStateAction());
@@ -65,25 +64,10 @@ export const LineDetailsPage = (): JSX.Element => {
     ? () => createRoute(line)
     : undefined;
 
-  const uniqueLineRouteLabels = pipe(
-    line?.line_routes,
-    (routes) => routes?.map((route) => route.label) || [],
-    (routeLabels) => uniq(routeLabels),
-  );
-
   const displayedRoutes =
     line?.line_routes?.filter((route) =>
       displayedRouteLabels?.includes(route.label),
     ) || [];
-
-  // If no route has been initially selected to display, show all line's routes
-  // Set the default value to query params if route labels query param doesn't exist
-  useEffect(() => {
-    if (!displayedRouteLabels && uniqueLineRouteLabels.length !== 0) {
-      setDisplayedRoutesToUrl(uniqueLineRouteLabels);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [uniqueLineRouteLabels]);
 
   return (
     <div>
